Made rdpRdp opaque
